_code/_tax_processing/requirements.txt
pandas>=2.0
requests>=2.31
PyYAML>=6.0
beautifulsoup4>=4.12
lxml>=5.0
pdfplumber>=0.11
python-dateutil>=2.9
